requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.27.2
slowapi==0.1.9
python-dotenv==1.0.1
elasticsearch==8.13.0
aiohttp==3.10.11
